java sources aliases
Include the java sources, which must be installed locally, as an alias.

Enables tooling, such as CIDER, to pick up the Java sources as an extra
dependency by including the alias

clojure -A:java-11-source ,,,

Paths to the Java src.zip file are from Ubuntu Linux distribution.  Modify the
path for your own system.