README: replace / with space in title
GitHub generates id's for titles based on names.  When a title contains a dash
character, that character is dropped.

The title Updating practicalli/clojure-deps-edn creates a hard to read id, so
changing the title to use a space

Updating the mini table of contents with the new link for the id.  This link is
currently broken

Resolves: #2